+++ client/tests/sdk/test_job.py
@@ -239,7 +239,7 @@ def evaluate_handler(*args, **kwargs): ...
         assert yaml_path.exists()
         jobs_info = load_yaml(yaml_path)
         assert jobs_info == {
-            "default": [
+            "evaluation": [
                 {
                     "cls_name": "",
                     "concurrency": 1,
@@ -252,7 +252,7 @@ def evaluate_handler(*args, **kwargs): ...
                         "ppl_batch_size": 1,
                     },
                     "func_name": "predict_handler",
-                    "job_name": "default",
+                    "job_name": "evaluation",
                     "module_name": self.module_name,
                     "name": "predict",
                     "needs": [],
@@ -265,7 +265,7 @@ def evaluate_handler(*args, **kwargs): ...
                     "extra_args": [],
                     "extra_kwargs": {"ppl_auto_log": True},
                     "func_name": "evaluate_handler",
-                    "job_name": "default",
+                    "job_name": "evaluation",
                     "module_name": self.module_name,
                     "name": "evaluate",
                     "needs": ["predict"],
@@ -275,7 +275,7 @@ def evaluate_handler(*args, **kwargs): ...
             ]
         }
 
-        steps = Step.get_steps_from_yaml("default", yaml_path)
+        steps = Step.get_steps_from_yaml("evaluation", yaml_path)
         assert len(steps) == 2
         assert steps[0].name == "predict"
         assert steps[1].name == "evaluate"
@@ -312,14 +312,14 @@ class MockHandler(PipelineHandler):
         assert yaml_path.exists()
         jobs_info = load_yaml(yaml_path)
         assert jobs_info == {
-            "default": [
+            "evaluation": [
                 {
                     "cls_name": "MockHandler",
                     "concurrency": 1,
                     "extra_args": [],
                     "extra_kwargs": {},
                     "func_name": "ppl",
-                    "job_name": "default",
+                    "job_name": "evaluation",
                     "module_name": self.module_name,
                     "name": "ppl",
                     "needs": [],
@@ -332,7 +332,7 @@ class MockHandler(PipelineHandler):
                     "extra_args": [],
                     "extra_kwargs": {},
                     "func_name": "cmp",
-                    "job_name": "default",
+                    "job_name": "evaluation",
                     "module_name": self.module_name,
                     "name": "cmp",
                     "needs": ["ppl"],
@@ -342,7 +342,7 @@ class MockHandler(PipelineHandler):
             ]
         }
 
-        steps = Step.get_steps_from_yaml("default", yaml_path)
+        steps = Step.get_steps_from_yaml("evaluation", yaml_path)
         context = Context(
             workdir=self.workdir,
             project="test",
@@ -381,7 +381,7 @@ class MockHandler:
         assert yaml_path.exists()
         jobs_info = load_yaml(yaml_path)
         assert jobs_info == {
-            "default": [
+            "evaluation": [
                 {
                     "cls_name": "MockHandler",
                     "concurrency": 1,
@@ -394,7 +394,7 @@ class MockHandler:
                         "ppl_batch_size": 1,
                     },
                     "func_name": "predict_handler",
-                    "job_name": "default",
+                    "job_name": "evaluation",
                     "module_name": self.module_name,
                     "name": "predict",
                     "needs": [],
@@ -407,7 +407,7 @@ class MockHandler:
                     "extra_args": [],
                     "extra_kwargs": {"ppl_auto_log": True},
                     "func_name": "evaluate_handler",
-                    "job_name": "default",
+                    "job_name": "evaluation",
                     "module_name": self.module_name,
                     "name": "evaluate",
                     "needs": ["predict"],
@@ -417,7 +417,7 @@ class MockHandler:
             ]
         }
 
-        steps = Step.get_steps_from_yaml("default", yaml_path)
+        steps = Step.get_steps_from_yaml("evaluation", yaml_path)
         results = Scheduler(
             project="test",
             version="test",
